:truck: refactor(server/main): Start splitting away migrations.
