+++ imports/app/atStartup.ts
@@ -1,0 +1,3 @@
+import {Meteor} from 'meteor/meteor';
+
+export default Meteor.startup;
